fixed issues related to checksum verification and applying committed data in replicas

+++ key-value-store/storage-server/src/main/java/com/bteshome/keyvaluestore/storage/states/OffsetState.java
@@ -5,6 +5,7 @@ import com.bteshome.keyvaluestore.common.LogPosition;
 import com.bteshome.keyvaluestore.common.MetadataCache;
 import com.bteshome.keyvaluestore.common.Utils;
 import com.bteshome.keyvaluestore.storage.common.ChecksumUtil;
+import com.bteshome.keyvaluestore.storage.common.CompressionUtil;
 import com.bteshome.keyvaluestore.storage.common.StorageServerException;
 import com.bteshome.keyvaluestore.storage.common.StorageSettings;
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -24,11 +25,11 @@ public class OffsetState {
     private final String table;
     private final int partition;
     private final String nodeId;
-    private final Map<String, LogPosition> replicaEndOffsets;
+    private LogPosition endOffset;
     private LogPosition committedOffset;
     private LogPosition previousLeaderEndOffset;
     private final ReentrantReadWriteLock lock;
-    private final String replicaEndOffsetsSnapshotFile;
+    private final String endOffsetSnapshotFile;
     private final String committedOffsetSnapshotFile;
     private final String previousLeaderEndOffsetFile;
 
@@ -36,42 +37,37 @@ public class OffsetState {
         this.table = table;
         this.partition = partition;
         this.nodeId = storageSettings.getNode().getId();
-        replicaEndOffsets = new ConcurrentHashMap<>();
         committedOffset = LogPosition.empty();
+        endOffset = LogPosition.empty();
         previousLeaderEndOffset = LogPosition.empty();
         lock = new ReentrantReadWriteLock(true);
-        replicaEndOffsetsSnapshotFile = "%s/%s-%s/replicaEndOffsets.ser".formatted(storageSettings.getNode().getStorageDir(), table, partition);
+        endOffsetSnapshotFile = "%s/%s-%s/endOffset.ser".formatted(storageSettings.getNode().getStorageDir(), table, partition);
         committedOffsetSnapshotFile = "%s/%s-%s/committedOffset.ser".formatted(storageSettings.getNode().getStorageDir(), table, partition);
         previousLeaderEndOffsetFile = "%s/%s-%s/previousLeaderEndOffset.ser".formatted(storageSettings.getNode().getStorageDir(), table, partition);
-        loadReplicaEndOffsetsFromSnapshot();
+        loadEndOffsetFromSnapshot();
         loadCommittedOffsetFromSnapshot();
         loadPreviousLeaderEndOffset();
     }
 
-    public Map<String, LogPosition> getReplicaEndOffsets() {
-        Map<String, LogPosition> copy;
+    public LogPosition getEndOffset() {
         try (AutoCloseableLock l = readLock()) {
-            copy = new HashMap<>(replicaEndOffsets);
-        }
-        return copy;
-    }
-
-    public LogPosition getReplicaEndOffset(String replicaNodeId) {
-        try (AutoCloseableLock l = readLock()) {
-            return replicaEndOffsets.getOrDefault(replicaNodeId, LogPosition.empty());
+            return endOffset;
         }
     }
 
-    public void setReplicaEndOffset(String replicaNodeId, LogPosition offset) {
-        try (AutoCloseableLock l = writeLock()) {
-            replicaEndOffsets.put(replicaNodeId, offset);
-        }
-    }
-
-    public void setReplicaEndOffsets(Map<String, LogPosition> offsets) {
-        for (Map.Entry<String, LogPosition> offset : offsets.entrySet()) {
-            if (!offset.getKey().equals(nodeId))
-                replicaEndOffsets.put(offset.getKey(), offset.getValue());
+    public void setEndOffset(LogPosition offset) {
+        BufferedWriter writer = Utils.createWriter(endOffsetSnapshotFile);
+        try (writer; AutoCloseableLock l = writeLock()) {
+            endOffset = offset;
+            // TODO - change to Java serialization once done testing
+            //writer.write(JsonSerDe.serialize(offset));
+            CompressionUtil.compressAndWrite(endOffsetSnapshotFile, offset);
+            ChecksumUtil.generateAndWrite(endOffsetSnapshotFile);
+            log.debug("Persisted end offset '{}' for table '{}' partition '{}'.", endOffset, table, partition);
+        } catch (IOException e) {
+            String errorMessage = "Error writing end offset for table '%s' partition '%s'.".formatted(table, partition);
+            log.error(errorMessage, e);
+            throw new StorageServerException(errorMessage, e);
         }
     }
 
@@ -86,7 +82,8 @@ public class OffsetState {
         try (writer; AutoCloseableLock l = writeLock()) {
             committedOffset = offset;
             // TODO - change to Java serialization once done testing
-            writer.write(JsonSerDe.serialize(offset));
+            //writer.write(JsonSerDe.serialize(offset));
+            CompressionUtil.compressAndWrite(committedOffsetSnapshotFile, offset);
             ChecksumUtil.generateAndWrite(committedOffsetSnapshotFile);
             log.debug("Persisted committed offset '{}' for table '{}' partition '{}'.", committedOffset, table, partition);
         } catch (IOException e) {
@@ -95,7 +92,6 @@ public class OffsetState {
             throw new StorageServerException(errorMessage, e);
         }
     }
-
 
     public LogPosition getPreviousLeaderEndOffset() {
         try (AutoCloseableLock l = readLock()) {
@@ -109,7 +105,9 @@ public class OffsetState {
         BufferedWriter writer = Utils.createWriter(previousLeaderEndOffsetFile);
         try (writer; AutoCloseableLock l = writeLock()) {
             previousLeaderEndOffset = offset;
-            writer.write(JsonSerDe.serialize(offset));
+            //writer.write(JsonSerDe.serialize(offset));
+            CompressionUtil.compressAndWrite(previousLeaderEndOffsetFile, offset);
+            ChecksumUtil.generateAndWrite(previousLeaderEndOffsetFile);
             log.debug("Persisted previous leader end offset '{}' for table '{}' partition '{}'.", previousLeaderEndOffset, table, partition);
         } catch (IOException e) {
             String errorMessage = "Error writing previous leader end offset for table '%s' partition '%s'.".formatted(table, partition);
@@ -120,10 +118,12 @@ public class OffsetState {
 
     public void clearPreviousLeaderEndOffset() {
         try (AutoCloseableLock l = writeLock()) {
-            previousLeaderEndOffset = LogPosition.empty();
-            Files.deleteIfExists(Path.of(previousLeaderEndOffsetFile));
-            Files.deleteIfExists(Path.of(previousLeaderEndOffsetFile + ".md5"));
-            log.debug("Deleted previous leader end offset file '{}' for table '{}' partition '{}'.", previousLeaderEndOffsetFile, table, partition);
+            if (!previousLeaderEndOffset.equals(LogPosition.empty())) {
+                previousLeaderEndOffset = LogPosition.empty();
+                Files.deleteIfExists(Path.of(previousLeaderEndOffsetFile));
+                Files.deleteIfExists(Path.of(previousLeaderEndOffsetFile + ".md5"));
+                log.debug("Deleted previous leader end offset file '{}' for table '{}' partition '{}'.", previousLeaderEndOffsetFile, table, partition);
+            }
         } catch (IOException e) {
             String errorMessage = "Error deleting previous leader end offset file for table '%s' partition '%s'.".formatted(table, partition);
             log.error(errorMessage, e);
@@ -131,38 +131,22 @@ public class OffsetState {
         }
     }
 
-    public void takeSnapshot() {
-        if (replicaEndOffsets.isEmpty())
+    private void loadEndOffsetFromSnapshot() {
+        if (Files.notExists(Path.of(endOffsetSnapshotFile)))
             return;
 
-        BufferedWriter writer = Utils.createWriter(replicaEndOffsetsSnapshotFile);
-        try (writer) {
-            // TODO - 1. change to Java serialization once done testing
-            writer.write(JsonSerDe.serialize(replicaEndOffsets));
-            ChecksumUtil.generateAndWrite(replicaEndOffsetsSnapshotFile);
-            log.debug("Took a snapshot of replica end offsets for table '{}' partition '{}'.", table, partition);
-        } catch (IOException e) {
-            String errorMessage = "Error taking a snapshot of replica end offsets for table '%s' partition '%s'.".formatted(table, partition);
-            log.error(errorMessage, e);
-            throw new StorageServerException(errorMessage, e);
-        }
-    }
+        ChecksumUtil.readAndVerify(endOffsetSnapshotFile);
 
-    private void loadReplicaEndOffsetsFromSnapshot() {
-        if (Files.notExists(Path.of(replicaEndOffsetsSnapshotFile)))
-            return;
-
-        ChecksumUtil.readAndVerify(replicaEndOffsetsSnapshotFile);
-
-        BufferedReader reader = Utils.createReader(replicaEndOffsetsSnapshotFile);
+        BufferedReader reader = Utils.createReader(endOffsetSnapshotFile);
         try (reader) {
             String line = reader.readLine();
             if (line != null) {
-                replicaEndOffsets.putAll(JsonSerDe.deserialize(line, new TypeReference<HashMap<String, LogPosition>>(){}));
-                log.debug("Loaded replica end offsets from snapshot for table '{}' partition '{}'.", table, partition);
+                //endOffset = JsonSerDe.deserialize(line, LogPosition.class);
+                endOffset = CompressionUtil.readAndDecompress(endOffsetSnapshotFile, LogPosition.class);
+                log.debug("Loaded end offset from snapshot for table '{}' partition '{}'.", table, partition);
             }
         } catch (IOException e) {
-            String errorMessage = "Error loading from a snapshot of replica end offsets for table '%s' partition '%s'.".formatted(table, partition);
+            String errorMessage = "Error loading from a snapshot of end offset for table '%s' partition '%s'.".formatted(table, partition);
             log.error(errorMessage, e);
             throw new StorageServerException(errorMessage, e);
         }
@@ -178,7 +162,8 @@ public class OffsetState {
         try (reader) {
             String line = reader.readLine();
             if (line != null) {
-                committedOffset = JsonSerDe.deserialize(line, LogPosition.class);
+                //committedOffset = JsonSerDe.deserialize(line, LogPosition.class);
+                committedOffset = CompressionUtil.readAndDecompress(committedOffsetSnapshotFile, LogPosition.class);
                 log.debug("Loaded replica committed offset from snapshot for table '{}' partition '{}'.", table, partition);
             }
         } catch (IOException e) {
@@ -198,7 +183,8 @@ public class OffsetState {
         try (reader) {
             String line = reader.readLine();
             if (line != null) {
-                previousLeaderEndOffset = JsonSerDe.deserialize(line, LogPosition.class);
+                //previousLeaderEndOffset = JsonSerDe.deserialize(line, LogPosition.class);
+                previousLeaderEndOffset = CompressionUtil.readAndDecompress(previousLeaderEndOffsetFile, LogPosition.class);
                 log.debug("Loaded previous leader end offset from snapshot for table '{}' partition '{}'.", table, partition);
             }
         } catch (IOException e) {
